backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
prophet==1.1.5
yfinance==0.2.28
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
celery==5.3.4
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1
ta==0.10.2
plotly==5.17.0
dash==2.14.2
